backend/requirements.txt
fastapi
uvicorn
boto3
pydantic-settings
python_dotenv
sqlalchemy
psycopg2-binary
redis